Update file structures and etc
