+++ .gitignore
@@ -149,6 +149,7 @@ cmake-build-debug/
 
 # IntelliJ
 out/
+src/out/
 
 # mpeltonen/sbt-idea plugin
 .idea_modules/
