--- infra/output.txt
@@ -1,45 +1,0 @@
-Collecting paramiko==2.4.1 (from -r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/3e/db/cb7b6656e0e7387637ce850689084dc0b94b44df31cc52e5fc5c2c4fd2c1/paramiko-2.4.1-py2.py3-none-any.whl
-Collecting pyasn1>=0.1.7 (from paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/a0/70/2c27740f08e477499ce19eefe05dbcae6f19fdc49e9e82ce4768be0643b9/pyasn1-0.4.3-py2.py3-none-any.whl
-Collecting cryptography>=1.5 (from paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/40/87/acdcf84ce6d25a7db1c113f4b9b614fd8d707b7ab56fbf17cf18cd26a627/cryptography-2.2.2-cp34-abi3-macosx_10_6_intel.whl
-Collecting bcrypt>=3.1.3 (from paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/7e/59/d48fd712941da1a5d6490964a37bb3de2e526965b6766273f6a7049ee590/bcrypt-3.1.4-cp36-cp36m-macosx_10_6_intel.whl
-Collecting pynacl>=1.0.1 (from paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/74/8e/a6c0d340972d9e2f1a405aaa3f2460950b4c0337f92db0291a4355974529/PyNaCl-1.2.1-cp36-cp36m-macosx_10_6_intel.whl
-Collecting six>=1.4.1 (from cryptography>=1.5->paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/67/4b/141a581104b1f6397bfa78ac9d43d8ad29a7ca43ea90a2d863fe3056e86a/six-1.11.0-py2.py3-none-any.whl
-Collecting cffi>=1.7; platform_python_implementation != "PyPy" (from cryptography>=1.5->paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/8e/be/40b1bc2c3221acdefeb9dab6773d43cda7543ed0d8c8df8768f05af2d01e/cffi-1.11.5-cp36-cp36m-macosx_10_6_intel.whl
-Collecting idna>=2.1 (from cryptography>=1.5->paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/27/cc/6dd9a3869f15c2edfab863b992838277279ce92663d334df9ecf5106f5c6/idna-2.6-py2.py3-none-any.whl
-Collecting asn1crypto>=0.21.0 (from cryptography>=1.5->paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-  Using cached https://files.pythonhosted.org/packages/ea/cd/35485615f45f30a510576f1a56d1e0a7ad7bd8ab5ed7cdc600ef7cd06222/asn1crypto-0.24.0-py2.py3-none-any.whl
-Collecting pycparser (from cffi>=1.7; platform_python_implementation != "PyPy"->cryptography>=1.5->paramiko==2.4.1->-r /Users/erengwa/myworkspace/github/chaos-lambda-latest/infra/../python_lib/requirements.txt (line 1))
-Installing collected packages: pyasn1, six, pycparser, cffi, idna, asn1crypto, cryptography, bcrypt, pynacl, paramiko
-Successfully installed asn1crypto-0.24.0 bcrypt-3.1.4 cffi-1.11.5 cryptography-2.2.2 idna-2.6 paramiko-2.4.1 pyasn1-0.4.3 pycparser-2.18 pynacl-1.2.1 six-1.11.0
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/__pycache__ already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/_cffi_backend.cpython-36m-darwin.so already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/asn1crypto already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/asn1crypto-0.24.0.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/bcrypt already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/bcrypt-3.1.4.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/cffi already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/cffi-1.11.5.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/cryptography already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/cryptography-2.2.2.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/idna already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/idna-2.6.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/nacl already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/paramiko already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/paramiko-2.4.1.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/pyasn1 already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/pyasn1-0.4.3.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/pycparser already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/pycparser-2.18.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/PyNaCl-1.2.1.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/six-1.11.0.dist-info already exists. Specify --upgrade to force replacement.
-Target directory /Users/erengwa/myworkspace/github/chaos-lambda-latest/python_lib/.tmp/six.py already exists. Specify --upgrade to force replacement.
-You are using pip version 9.0.1, however version 10.0.1 is available.
-You should consider upgrading via the 'pip install --upgrade pip' command.
