--- python_lib/requirements.txt
@@ -1,1 +1,0 @@
-paramiko==2.4.1
