+++ src/chaos.py
@@ -1,7 +1,8 @@
 import os
 import random
-from time import strftime, gmtime
 import boto3
+from tasks import *
+from helper import *
 
 REGIONS_VAIRABLE_NAME = "regions"
 ASG_GROUP_NAME = "asg_group_name"
@@ -103,25 +104,6 @@ def run_chaos_each_account_region(account, instances, region):
         results.append(result)
     return results
 
-# def terminate_no_point_of_return(account, instance, region):
-#     ec2 = assumRole(account, "ec2", region)
-#     ec2.terminate_instances(InstanceIds=[instance[1]])
-#     result = "TERMINATE", instance[1], "from", instance[0], "in", region
-#     printlog(result)
-#     return result
-
-
-# def terminate_dry_run(instance, region):
-#     result = "Terminate [DRY-RUN] {} from {} in {}".format(instance[1],
-#                                                            instance[0], region)
-#     printlog(result)
-#     return result
-
-
-def printlog(*args):
-    current = strftime("%Y-%m-%d %H:%M:%SZ", gmtime())
-    print(current, *args)
-
 
 def string_to_bool(s):
     if s.lower() in ["true", "t", "yes", "yeah", "yup", "y", "certainly",
@@ -184,52 +166,6 @@ def assumRole(account, service, region):
     return client
 
 """
-Chaos Tasks
-"""
-def calling_tasks_random(account, i, region, dryrun=True):
-    random_task = random.randint(0, len(TASKS)-1)
-    task_to_run, descritpion = TASKS[random_task]
-    return task_to_run(account, i, region, dryrun) 
-
-
-def terminate_instance_worker(account, instance, region, dryrun=True):
-    if dryrun:
-        result = "Terminate [DRY-RUN] {} from {} in {}".format(instance[1],
-                                                           instance[0], region)
-        printlog(result)
-        return result
-    else:    
-        ec2 = assumRole(account, "ec2", region)
-        ec2.terminate_instances(InstanceIds=[instance[1]])
-        result = "TERMINATE", instance[1], "from", instance[0], "in", region
-        printlog(result)
-        return result
-
-
-def max_cpu_worker(account, instance, region, dryrun=True):
-    if dryrun:
-        result = "Max out CPU [DRY-RUN] {} from {} in {}".format(instance[1],
-                                                           instance[0], region)
-        printlog(result)
-        return result
-    else:
-        result = "Max out CPU {} from {} in {}".format(instance[1],
-                                                           instance[0], region)
-        ssm = assumRole(account, "ssm", region)
-        resp = ssm.send_command(
-            DocumentName="AWS-RunShellScript",
-            Parameters={'commands': ["cat << EOF > /tmp/infiniteburn.sh","#!/bin/bash","while true;"," do openssl speed;","done","EOF","","# 32 parallel 100% CPU tasks should hit even the biggest EC2 instances","for i in {1..32}","do"," nohup /bin/bash /tmp/infiniteburn.sh > /dev/null 2>&1 &","done"]},
-            InstanceIds=[instance[1]]
-        )
-        printlog(resp)
-        return result
-
-TASKS = [
-    (terminate_instance_worker, "Terminate instances"),
-    (max_cpu_worker, "Max out CPU")
-]
-
-"""
 Main Handler function
 """
 def handler(event, context):
@@ -242,6 +178,6 @@ def handler(event, context):
     result = run_chaos(accounts, regions, global_prob)
     alert(result)
 
-# if __name__ == '__main__':
-#     # calling_tasks_random("152303423357", ["asg","i-06d3fe93310a66d69"], "us-east-1", dryrun=True)
-#     # max_cpu_worker("152303423357", ["asg","i-06d3fe93310a66d69"], "us-east-1", dryrun=False)
+if __name__ == '__main__':
+    calling_tasks_random("152303423357", ["asg","i-06d3fe93310a66d69"], "us-east-1", dryrun=True)
+    # max_cpu_worker("152303423357", ["asg","i-06d3fe93310a66d69"], "us-east-1", dryrun=False)
